+++ src/components/Pages/CreateCompo/CreateCompo.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import {
     ConfigHeader,
     PlayerNumber,
+    roles,
     RolesChoice,
     useAppDispatch,
     useAppSelector,
@@ -20,6 +21,28 @@ export const CreateCompo: React.FC = () => {
         dispatch(werewolfActions.changePlayerNumber(newValue));
     };
 
+    const {composition} = gameData;
+
+    const roleSelected = composition.reduce((a, b) => {
+        return a + b.quantity;
+    }, 0);
+
+    let roleNeeded = gameData.playerNumber;
+
+    if (composition.find((el) => {
+        return el.quantity === 1 && el.roleName === "voleur";
+    })) {
+        roleNeeded += 2;
+    }
+
+    if (composition.find((el) => {
+        return el.quantity === 1 && el.roleName === "comedien";
+    })) {
+        roleNeeded += 3;
+    }
+
+    const maxPLayer = roles.reduce((a, b) => a + b.max, 0) - 2;
+
     return (
         <div className="create-compo-box">
             <ConfigHeader
@@ -32,12 +55,20 @@ export const CreateCompo: React.FC = () => {
             </p>
             <PlayerNumber
                 changePlayerNumber={changePlayerNumber}
-                max={47}
+                max={maxPLayer}
                 min={4}
                 value={gameData.playerNumber}
             />
             <p className="config-section-title">
-                Choisir la composition
+                <span>
+                    Choisir la composition (
+                </span>
+                <span className="role-number-display">
+                    {roleSelected + "/" + roleNeeded}
+                </span>
+                <span>
+                    )
+                </span>
             </p>
             <RolesChoice />
         </div>